add Timer2_PWM

+++ inc/delay.h
@@ -3,8 +3,8 @@
 #ifndef __DELAY_H__
 #define __DELAY_H__
 
-void delay_nms(int ms);
-void delay_nus(int us);            // fixed-offset:56us
+void delay_nms(short ms);
+void delay_nus(short us);            // fixed-offset:56us
 void delay_100nop();
 void delay_5nop();
